made ShowAvatar in settings it's own component
