+++ src/components/Settings/Settings.tsx
@@ -6,51 +6,28 @@ import FormControlLabel from '@mui/material/FormControlLabel'
 import Box from '@mui/material/Box'
 import MenuItem from '@mui/material/MenuItem'
 import {
-  setAvatarVisibility,
   selectIsSettingsOpen,
   setIsSettingsOpen,
-  setShowAvatar,
   setEmailFetchSize,
   selectEmailListSize,
 } from '../../Store/utilsSlice'
 import { useAppDispatch, useAppSelector } from '../../Store/hooks'
 import * as S from './SettingsStyles'
 import * as global from '../../constants/globalConstants'
+import ShowAvatar from './SettingsOptions/ShowAvatar/ShowAvatar'
 
 const handleClose = (dispatch: Function) => dispatch(setIsSettingsOpen(false))
 
 const SETTINGS = 'Settings'
-const showAvatarLocalStorage = localStorage.getItem('showAvatar')
 
 const Settings = () => {
   const dispatch = useAppDispatch()
   const isSettingsOpen = useAppSelector(selectIsSettingsOpen)
-  const avatarVisible = useAppSelector(setAvatarVisibility)
   const fetchCount = useAppSelector(selectEmailListSize)
 
   const handleEmailListSizeChange = (e: any) => {
     localStorage.setItem('fetchSize', e.target.value)
     dispatch(setEmailFetchSize(e.target.value))
-  }
-
-  useEffect(() => {
-    if (
-      !showAvatarLocalStorage ||
-      (showAvatarLocalStorage !== 'true' && showAvatarLocalStorage !== 'false')
-    ) {
-      localStorage.setItem('showAvatar', 'true')
-      dispatch(setShowAvatar(true))
-    }
-  }, [])
-
-  const switchAvatarView = () => {
-    if (localStorage.getItem('showAvatar') === 'true') {
-      localStorage.setItem('showAvatar', 'false')
-      dispatch(setShowAvatar(false))
-    } else {
-      localStorage.setItem('showAvatar', 'true')
-      dispatch(setShowAvatar(true))
-    }
   }
 
   return (
@@ -64,16 +41,7 @@ const Settings = () => {
         <S.SettingsHeader>{SETTINGS}</S.SettingsHeader>
         <S.SettingsContainer>
           <FormGroup>
-            <FormControlLabel
-              label="Do you want to see Avatars?"
-              control={
-                <Switch
-                  onClick={() => switchAvatarView()}
-                  checked={avatarVisible}
-                  color="secondary"
-                />
-              }
-            />
+            <ShowAvatar />
 
             <FormControlLabel
               label="Emails Fetched at a time"
